New functionality for fastrpc.
Include a bug fixing.

+++ include/fastcdr/FastCdr.h
@@ -60,7 +60,7 @@ namespace eprosima
          * @param numBytes The number of bytes that will be jumped.
          * @return True is returned when the jump operation works successfully. Otherwise, false is returned.
          */
-        bool jump(uint32_t numBytes);
+        bool jump(size_t numBytes);
 
         /*!
 		 * @brief This function resets the current position in the buffer to the begining.
